requirements.txt
requests>=2.28
beautifulsoup4>=4.12
python-dateutil>=2.8
